+++ changelog.md
@@ -1,2 +1,2 @@
 ### Chunk Loaders 1.2.9
-- Initial release of Chunk Loaders for Minecraft 1.21.9 & 1.21.10
+- Initial release of Chunk Loaders for Minecraft 1.21.11